simplify data ingestion process for query testing
